refactor: use nano-queries
