--- src/utils/SQLBuilder/SetExpression.ts
@@ -1,22 +1,0 @@
-import { RawQuery } from './core/RawQuery';
-import { GroupExpression } from './GroupExpression';
-import { IQuery, QuerySegment } from './types';
-import { QueryConstructor } from './utils/QueryConstructor';
-
-export class SetExpression extends RawQuery implements IQuery {
-	public withParenthesis() {
-		return new GroupExpression(this);
-	}
-
-	public exportQuery(): QuerySegment[] {
-		const query = new QueryConstructor();
-
-		super.exportQuery().forEach((item, index) => {
-			const preparedItem =
-				item instanceof SetExpression ? item.withParenthesis() : item;
-			query.raw(index > 0 ? ',' : undefined, preparedItem);
-		});
-
-		return query.exportQuery();
-	}
-}
