--- src/utils/SQLBuilder/core/PreparedValue.ts
@@ -1,3 +1,0 @@
-import { RawValue } from './RawValue';
-
-export class PreparedValue extends RawValue {}
