--- src/utils/SQLBuilder/core/RawValue.ts
@@ -1,10 +1,0 @@
-export class RawValue {
-	protected readonly value;
-	constructor(value: string | number | null) {
-		this.value = value;
-	}
-
-	public getValue = () => {
-		return this.value;
-	};
-}
