--- src/utils/SQLBuilder/types.ts
@@ -1,27 +1,0 @@
-import { PreparedValue } from './core/PreparedValue';
-import { RawQuery } from './core/RawQuery';
-import { RawValue } from './core/RawValue';
-
-export type QueryBindings = string | number | null;
-export type QuerySegment = RawValue | RawQuery | PreparedValue;
-export type QueryParameter = QuerySegment | QueryBindings;
-
-export type RawQueryParameter = QueryParameter | undefined;
-
-export interface IQuery {
-	/**
-	 * Returns query segments number
-	 */
-	size(): number;
-
-	/**
-	 * Returns final query that may be preprocessed
-	 * Returned query will be used to compile SQL
-	 */
-	exportQuery(): QuerySegment[];
-
-	/**
-	 * Compile query to SQL string and bindings
-	 */
-	toSQL(): { sql: string; bindings: QueryBindings[] };
-}
